face_recognition/embedder.py
"""
Face Embedder Module (deepface-onnx)
====================================

This module generates face embeddings using FaceNet512 from deepface-onnx.
Provides 512-dimensional embeddings for face recognition.

VALIDATED:
- Same person similarity: 0.82-0.86
- Different person similarity: 0.03-0.09
- Model size: 94MB (FaceNet512)

USAGE:
------
    from face_recognition.embedder import FaceEmbedder
    
    embedder = FaceEmbedder()
    embedding = embedder.get_embedding(aligned_face)
    # embedding = numpy array of shape (512,)
"""

import cv2
import numpy as np
import os
import sys
from pathlib import Path
from typing import Optional, List

# Add deepface-onnx to path
_lib_path = Path(__file__).parent.parent / "lib" / "deepface_onnx"
if str(_lib_path) not in sys.path:
    sys.path.insert(0, str(_lib_path))

from .config import Config


class FaceEmbedder:
    """
    Face embedder using FaceNet512 from deepface-onnx.
    
    FaceNet512 produces 512-dimensional embeddings that are:
    - High similarity (0.8+) for same person
    - Low similarity (0.1-) for different people
    
    Model is automatically downloaded on first use (~94MB).
    """
    
    # Model configuration
    MODEL_NAME = "facenet512"
    INPUT_SIZE = (160, 160)  # FaceNet512 input size
    EMBEDDING_DIM = 512
    
    def __init__(self):
        """Initialize the face embedder."""
        self.model_path = _lib_path / "models" / f"{self.MODEL_NAME}.onnx"
        self.session = None
        self.input_name = None
        self.output_name = None
        
        # Load model
        self._load_model()
    
    def _load_model(self):
        """Load the FaceNet512 ONNX model."""
        # Change to lib directory for proper model download path
        original_cwd = os.getcwd()
        os.chdir(str(_lib_path))
        
        try:
            # Import and use deepface-onnx's DeepFace class
            from utils import DeepFace
            
            # This will download the model if not present
            df = DeepFace(model_name=self.MODEL_NAME)
            
            # Store the session for direct use
            self.session = df.model
            self.input_name = df.input_name
            self.output_name = df.output_name
            
            print(f"✓ Face embedder loaded: FaceNet512 ({self.EMBEDDING_DIM} dimensions)")
        finally:
            os.chdir(original_cwd)
    
    def preprocess(self, face_image: np.ndarray) -> np.ndarray:
        """
        Preprocess a face image for the embedder.
        
        Steps:
        1. Resize to model input size (160x160)
        2. Normalize to [0, 1] range
        3. Add batch dimension
        
        Args:
            face_image: Aligned face image (BGR, any size)
        
        Returns:
            Preprocessed image ready for inference
        """
        # Resize to expected input size
        if face_image.shape[:2] != self.INPUT_SIZE:
            face_image = cv2.resize(face_image, self.INPUT_SIZE)
        
        # Add batch dimension and convert to float
        face_image = np.expand_dims(face_image, axis=0)
        
        # Normalize to [0, 1]
        if face_image.max() > 1:
            face_image = face_image.astype(np.float32) / 255.0
        
        return face_image.astype(np.float32)
    
    def get_embedding(self, face_image: np.ndarray) -> np.ndarray:
        """
        Generate embedding for a single face image.
        
        Args:
            face_image: Aligned face image (BGR format)
        
        Returns:
            512-dimensional embedding vector (normalized)
        """
        # Preprocess
        input_data = self.preprocess(face_image)
        
        # Run inference
        outputs = self.session.run(
            [self.output_name], 
            {self.input_name: input_data}
        )
        
        # Get embedding (first output, first batch item)
        embedding = outputs[0][0]
        
        # Normalize the embedding (required for cosine similarity)
        embedding = embedding / np.linalg.norm(embedding)
        
        return embedding.astype(np.float32)
    
    def get_embeddings_batch(self, face_images: List[np.ndarray]) -> np.ndarray:
        """
        Generate embeddings for multiple faces in a batch.
        
        More efficient than calling get_embedding() in a loop.
        
        Args:
            face_images: List of aligned face images
        
        Returns:
            Array of embeddings, shape (N, 512)
        """
        if not face_images:
            return np.array([])
        
        # Preprocess all faces
        batch = np.vstack([self.preprocess(face) for face in face_images])
        
        # Run inference on batch
        outputs = self.session.run(
            [self.output_name],
            {self.input_name: batch}
        )
        
        # Get embeddings
        embeddings = outputs[0]
        
        # Normalize each embedding
        norms = np.linalg.norm(embeddings, axis=1, keepdims=True)
        embeddings = embeddings / norms
        
        return embeddings.astype(np.float32)


def cosine_similarity(embedding1: np.ndarray, embedding2: np.ndarray) -> float:
    """
    Compute cosine similarity between two embeddings.
    
    If embeddings are normalized, this is just the dot product.
    
    Args:
        embedding1: First embedding (512-dim)
        embedding2: Second embedding (512-dim)
    
    Returns:
        Similarity score between -1 and 1 (higher = more similar)
    """
    # Normalize if not already normalized
    e1 = embedding1 / np.linalg.norm(embedding1)
    e2 = embedding2 / np.linalg.norm(embedding2)
    
    # Dot product of normalized vectors = cosine similarity
    return float(np.dot(e1, e2))


# Singleton instance for reuse
_embedder_instance = None

def get_embedder():
    """
    Get or create the singleton FaceNet512 embedder instance.
    
    Returns the same embedder for both:
    - Ingestion (creating FAISS index)
    - Query (user uploads/selfies)
    """
    global _embedder_instance
    
    if _embedder_instance is None:
        print("🔷 Using FaceNet512 embedder (99.65% LFW)")
        _embedder_instance = FaceEmbedder()
    
    return _embedder_instance


def reset_embedder():
    """Reset the singleton instance (useful for switching models in testing)."""
    global _embedder_instance
    _embedder_instance = None

